requirements.txt
numpy==1.26.4
scipy==1.12.0
scikit-learn==1.4.2
faiss-cpu==1.8.0
networkx==3.2.1
pydantic==2.6.3
pyyaml==6.0.1
typer==0.12.3
pytest==7.4.4
tqdm==4.66.2
